# Api/api/docs/swagger.yaml
definitions:
...
  model.ChangePassRes:
    properties:
      id:
        type: string
      new_pass:
        type: string
    type: object
  model.ChengePass:
    properties:
      id:
        type: string
      new_pass:
        type: string
      old_pass:
        type: string
      very_new:
        type: string
    type: object
...
  model.CreateUser:
    properties:
      address:
        minLength: 7
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        minLength: 5
        type: string
      phone:
        minLength: 5
        type: string
      postalcode:
        type: string
      role:
        type: string
      username:
        minLength: 4
        type: string
    required:
    - address
    - email
    - first_name
    - gender
    - last_name
    - password
    - phone
    - postalcode
    - role
    - username
    type: object
  model.Error:
    properties:
      message:
        type: string
    type: object
...
  model.Post:
    properties:
      author_id:
        type: string
      body:
        type: string
      brand_id:
        type: string
      category_id:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      gen:
        type: string
      id:
        type: string
      price:
        type: string
      product_type:
        type: string
      rating:
        type: string
      size:
        items:
          type: string
        type: array
      stars:
        maxLength: 5
        minLength: 0
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - brand_id
    - category_id
    - gen
    - stars
    type: object
...
  model.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model.Error'
    type: object
...
  model.User:
    properties:
      Posts:
        items:
          $ref: '#/definitions/model.Post'
        type: array
      address:
        type: string
      code:
        type: string
      color:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      postalcode:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.Users:
    properties:
      Users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  v1.Success:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/user/changePass:
    put:
      consumes:
      - application/json
      description: This API for changing user password
      parameters:
      - description: ChangePassword
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.ChengePass'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChangePassRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: Change User Password
      tags:
      - user
  /v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting user by ID
      parameters:
      - description: User_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Success'
      security:
      - BearerAuth: []
      summary: Delete User By Id
      tags:
      - user
  /v1/user/getUserbyId/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting user by id
      parameters:
      - description: User_ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Success'
      security:
      - BearerAuth: []
      summary: GetUserById
      tags:
      - user
  /v1/users:
    post:
      consumes:
      - application/json
      description: This API for creating a new user
      parameters:
      - description: CreateUser
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Success'
      summary: CreateUser
      tags:
      - user
    put:
      consumes:
      - application/json
      description: This API for updating a new user
      parameters:
      - description: UpdateUser
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Success'
      security:
      - BearerAuth: []
      summary: UpdateUser
      tags:
      - user
...
  /v1/users/getAll:
    get:
      consumes:
      - application/json
      description: This API for getting all Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Users'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Success'
      summary: GetAllUsers
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"